+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@ Changes:
 
   * Allowed assigning a barcode to loose storage
   * Record the user who sets QC status on samples, libraries, library aliquots, and run-libraries
+  * Improved performance of Edit Project page
   * Fixed RNA Library Preparation spreadsheet download and Pinery-MISO excluding some QCs if QC type
     name doesn't exactly match the original spec.
   * Fixed error displaying/validating QC type when editing QC with an archived QC type